Move fixture class and factories to mr_test header.
